Moved export from code-behind to view model.
