+++ MetaMorpheus/GUI/MetaDraw/MetaDraw.xaml
@@ -496,7 +496,7 @@
             </TabItem>
             
             <!-- Chimera Analysis tab -->
-            <TabItem Header="Chimera Analysis" Background="DarkGray" x:Name="ChimeraTab">
+            <TabItem Header="Chimera Annotation" Background="DarkGray" x:Name="ChimeraTab">
                 <local:ChimeraAnalysisTabView x:Name="ChimeraAnalysisTabView" />
             </TabItem>
         </TabControl>
